shared/pyproject.toml
[project]
name = "ai-study-shared"
version = "0.1.0"
description = "Shared utilities for the 14-day LLM bootcamp"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # LLM providers
    "openai>=1.52.0",
    "anthropic>=0.39.0",
    "google-genai>=0.3.0",

    # Local
    "ollama>=0.4.0",

    # Structured / validation
    "pydantic>=2.9.0",
    "instructor>=1.7.0",

    # Tokens / cost
    "tiktoken>=0.8.0",

    # HTTP / env
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",

    # Vector / embedding
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    "rank-bm25>=0.2.2",

    # Doc parsing
    "pypdf>=5.1.0",

    # LLM frameworks (Day 7+)
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.50",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "llama-index>=0.12.0",

    # Eval
    "ragas>=0.2.0",
    "datasets>=3.1.0",

    # Observability
    "langfuse>=2.55.0",

    # MCP (Day 11)
    "mcp>=1.1.0",

    # Web / UI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",

    # Tools (Day 5)
    "tavily-python>=0.5.0",

    # Logging
    "loguru>=0.7.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ipython>=8.28.0",
    "jupyterlab>=4.3.0",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "TID"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_study"]
